LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-name-inline-001-expected.txt

12 123

FAIL .target 1 assert_equals:
<span class="target" data-expected-width="30"></span>
width expected 30 but got 20
FAIL .target 2 assert_equals:
<span class="target" data-expected-width="30"></span>
width expected 30 but got 20
PASS .target 3
PASS .target 4
PASS .target 5
PASS .target 6
PASS .target 7
